Add files via upload
MongoDB Mini-Project: Vehicle Rental Agency Database
This project consists of an API written in Java to communicate with a MongoDB database representing a vehicle rental business. The task is to model the entities for vehicles, clients, administrators, ... and all the interactions between them. The details of the assignment are in the Mini projet.pdf file (written in French)

This project is carried out as part of a course called Advanced Databases for the 3rd year engineering students at the Higher Institute of Computer Science, Tunis (ISI)

  If you found this helpful, consider helping me by sending some Bitcoins to the following address. Thank you 😊

15m4Prjz4WLSsCgBcZTa62cUetiwo6sZKg